requirements.txt
requests>=2.31.0,<3.0
holidays>=0.44,<1.0
google-auth>=2.38.0,<3.0
